Make it usable for building outside of bincrafters
The changes make it possible to override all the multi-packager variables from the usual conan env vars so that one can set up custom local one. For example to reroute building on CI to a personal account. It also adds parsing of a `BUILD` env var that contains space delimited env vars to apply during a build to allow for shorthand CI build variations. Finally, it uses "*:shared" for the shared option as packages might not have that option specifically and using the wildcard will allow building those without errors.