Correct copyright year
This corrects my former introduced mistake. For all other repositories all years in which files got modified should be listed. However, since this is the templates repository and gets used for creating new repositories it would be wrong to also list a year from the past in a newly created repository.